refactor: bring all handlers together in EventHandler module
