+++ src/index.js
@@ -1,9 +1,3 @@
 import {VirtualKeyboard} from '@models/VirtualKeyboard.js';
 
 new VirtualKeyboard();
-
-// setTimeout(() => {
-//   console.log('it\'s working');
-//   document.execCommand('forwardDelete', false, null);
-//
-// }, 5000);
