+++ src/models/keyboard-template.js
@@ -1,0 +1,693 @@
+export const keyboardTemplate = () => {
+  return `
+    <div class="keyboard">
+      <div class="row">
+        <div class="key" data-key-code="Backquote">
+          <div class="key__ru">
+            <div class="key__default">ё</div>
+            <div class="key__shift">Ё</div>
+            <div class="key__caps-lock">Ё</div>
+            <div class="key__shift-caps-lock">ё</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">\`</div>
+            <div class="key__shift">~</div>
+            <div class="key__caps-lock">\`</div>
+            <div class="key__shift-caps-lock">~</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit1">
+          <div class="key__ru">
+            <div class="key__default">1</div>
+            <div class="key__shift">!</div>
+            <div class="key__caps-lock">1</div>
+            <div class="key__shift-caps-lock">!</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">1</div>
+            <div class="key__shift">!</div>
+            <div class="key__caps-lock">1</div>
+            <div class="key__shift-caps-lock">!</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit2">
+          <div class="key__ru">
+            <div class="key__default">2</div>
+            <div class="key__shift">"</div>
+            <div class="key__caps-lock">2</div>
+            <div class="key__shift-caps-lock">"</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">2</div>
+            <div class="key__shift">@</div>
+            <div class="key__caps-lock">2</div>
+            <div class="key__shift-caps-lock">@</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit3">
+          <div class="key__ru">
+            <div class="key__default">3</div>
+            <div class="key__shift">№</div>
+            <div class="key__caps-lock">3</div>
+            <div class="key__shift-caps-lock">№</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">3</div>
+            <div class="key__shift">#</div>
+            <div class="key__caps-lock">3</div>
+            <div class="key__shift-caps-lock">#</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit4">
+          <div class="key__ru">
+            <div class="key__default">4</div>
+            <div class="key__shift">;</div>
+            <div class="key__caps-lock">4</div>
+            <div class="key__shift-caps-lock">;</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">4</div>
+            <div class="key__shift">$</div>
+            <div class="key__caps-lock">4</div>
+            <div class="key__shift-caps-lock">$</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit5">
+          <div class="key__ru">
+            <div class="key__default">5</div>
+            <div class="key__shift">%</div>
+            <div class="key__caps-lock">5</div>
+            <div class="key__shift-caps-lock">%</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">5</div>
+            <div class="key__shift">%</div>
+            <div class="key__caps-lock">5</div>
+            <div class="key__shift-caps-lock">%</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit6">
+          <div class="key__ru">
+            <div class="key__default">6</div>
+            <div class="key__shift">:</div>
+            <div class="key__caps-lock">6</div>
+            <div class="key__shift-caps-lock">:</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">6</div>
+            <div class="key__shift">^</div>
+            <div class="key__caps-lock">6</div>
+            <div class="key__shift-caps-lock">^</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit7">
+          <div class="key__ru">
+            <div class="key__default">7</div>
+            <div class="key__shift">?</div>
+            <div class="key__caps-lock">7</div>
+            <div class="key__shift-caps-lock">?</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">7</div>
+            <div class="key__shift">&amp;</div>
+            <div class="key__caps-lock">7</div>
+            <div class="key__shift-caps-lock">&amp;</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit8">
+          <div class="key__ru">
+            <div class="key__default">8</div>
+            <div class="key__shift">*</div>
+            <div class="key__caps-lock">8</div>
+            <div class="key__shift-caps-lock">*</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">8</div>
+            <div class="key__shift">*</div>
+            <div class="key__caps-lock">8</div>
+            <div class="key__shift-caps-lock">*</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit9">
+          <div class="key__ru">
+            <div class="key__default">9</div>
+            <div class="key__shift">(</div>
+            <div class="key__caps-lock">9</div>
+            <div class="key__shift-caps-lock">(</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">9</div>
+            <div class="key__shift">(</div>
+            <div class="key__caps-lock">9</div>
+            <div class="key__shift-caps-lock">(</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Digit0">
+          <div class="key__ru">
+            <div class="key__default">0</div>
+            <div class="key__shift">)</div>
+            <div class="key__caps-lock">0</div>
+            <div class="key__shift-caps-lock">)</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">0</div>
+            <div class="key__shift">)</div>
+            <div class="key__caps-lock">0</div>
+            <div class="key__shift-caps-lock">)</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Minus">
+          <div class="key__ru">
+            <div class="key__default">-</div>
+            <div class="key__shift">_</div>
+            <div class="key__caps-lock">-</div>
+            <div class="key__shift-caps-lock">_</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">-</div>
+            <div class="key__shift">_</div>
+            <div class="key__caps-lock">-</div>
+            <div class="key__shift-caps-lock">_</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Equal">
+          <div class="key__ru">
+            <div class="key__default">=</div>
+            <div class="key__shift">+</div>
+            <div class="key__caps-lock">=</div>
+            <div class="key__shift-caps-lock">+</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">=</div>
+            <div class="key__shift">+</div>
+            <div class="key__caps-lock">=</div>
+            <div class="key__shift-caps-lock">+</div>
+          </div>
+        </div>
+        <div class="key key--expand key--word-right" data-key-code="Backspace">backspace</div>
+      </div>
+      <div class="row">
+        <div class="key key--expand key--word-left" data-key-code="Tab">tab</div>
+        <div class="key" data-key-code="KeyQ">
+          <div class="key__ru">
+            <div class="key__default">й</div>
+            <div class="key__shift">Й</div>
+            <div class="key__caps-lock">Й</div>
+            <div class="key__shift-caps-lock">й</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">q</div>
+            <div class="key__shift">Q</div>
+            <div class="key__caps-lock">Q</div>
+            <div class="key__shift-caps-lock">q</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyW">
+          <div class="key__ru">
+            <div class="key__default">ц</div>
+            <div class="key__shift">Ц</div>
+            <div class="key__caps-lock">Ц</div>
+            <div class="key__shift-caps-lock">ц</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">w</div>
+            <div class="key__shift">W</div>
+            <div class="key__caps-lock">W</div>
+            <div class="key__shift-caps-lock">w</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyE">
+          <div class="key__ru">
+            <div class="key__default">у</div>
+            <div class="key__shift">У</div>
+            <div class="key__caps-lock">У</div>
+            <div class="key__shift-caps-lock">у</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">e</div>
+            <div class="key__shift">E</div>
+            <div class="key__caps-lock">E</div>
+            <div class="key__shift-caps-lock">e</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyR">
+          <div class="key__ru">
+            <div class="key__default">к</div>
+            <div class="key__shift">К</div>
+            <div class="key__caps-lock">К</div>
+            <div class="key__shift-caps-lock">к</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">r</div>
+            <div class="key__shift">R</div>
+            <div class="key__caps-lock">R</div>
+            <div class="key__shift-caps-lock">r</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyT">
+          <div class="key__ru">
+            <div class="key__default">е</div>
+            <div class="key__shift">Е</div>
+            <div class="key__caps-lock">Е</div>
+            <div class="key__shift-caps-lock">е</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">t</div>
+            <div class="key__shift">T</div>
+            <div class="key__caps-lock">T</div>
+            <div class="key__shift-caps-lock">t</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyY">
+          <div class="key__ru">
+            <div class="key__default">н</div>
+            <div class="key__shift">Н</div>
+            <div class="key__caps-lock">Н</div>
+            <div class="key__shift-caps-lock">н</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">y</div>
+            <div class="key__shift">Y</div>
+            <div class="key__caps-lock">Y</div>
+            <div class="key__shift-caps-lock">y</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyU">
+          <div class="key__ru">
+            <div class="key__default">г</div>
+            <div class="key__shift">Г</div>
+            <div class="key__caps-lock">Г</div>
+            <div class="key__shift-caps-lock">г</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">u</div>
+            <div class="key__shift">U</div>
+            <div class="key__caps-lock">U</div>
+            <div class="key__shift-caps-lock">u</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyI">
+          <div class="key__ru">
+            <div class="key__default">ш</div>
+            <div class="key__shift">Ш</div>
+            <div class="key__caps-lock">Ш</div>
+            <div class="key__shift-caps-lock">ш</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">i</div>
+            <div class="key__shift">I</div>
+            <div class="key__caps-lock">I</div>
+            <div class="key__shift-caps-lock">i</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyO">
+          <div class="key__ru">
+            <div class="key__default">щ</div>
+            <div class="key__shift">Щ</div>
+            <div class="key__caps-lock">Щ</div>
+            <div class="key__shift-caps-lock">щ</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">o</div>
+            <div class="key__shift">O</div>
+            <div class="key__caps-lock">O</div>
+            <div class="key__shift-caps-lock">o</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyP">
+          <div class="key__ru">
+            <div class="key__default">з</div>
+            <div class="key__shift">З</div>
+            <div class="key__caps-lock">З</div>
+            <div class="key__shift-caps-lock">з</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">p</div>
+            <div class="key__shift">P</div>
+            <div class="key__caps-lock">P</div>
+            <div class="key__shift-caps-lock">p</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="BracketLeft">
+          <div class="key__ru">
+            <div class="key__default">х</div>
+            <div class="key__shift">Х</div>
+            <div class="key__caps-lock">Х</div>
+            <div class="key__shift-caps-lock">х</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">[</div>
+            <div class="key__shift">{</div>
+            <div class="key__caps-lock">[</div>
+            <div class="key__shift-caps-lock">{</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="BracketRight">
+          <div class="key__ru">
+            <div class="key__default">ъ</div>
+            <div class="key__shift">Ъ</div>
+            <div class="key__caps-lock">Ъ</div>
+            <div class="key__shift-caps-lock">ъ</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">]</div>
+            <div class="key__shift">}</div>
+            <div class="key__caps-lock">]</div>
+            <div class="key__shift-caps-lock">}</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Backslash">
+          <div class="key__ru">
+            <div class="key__default">\\</div>
+            <div class="key__shift">/</div>
+            <div class="key__caps-lock">\\</div>
+            <div class="key__shift-caps-lock">/</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">\\</div>
+            <div class="key__shift">|</div>
+            <div class="key__caps-lock">\\</div>
+            <div class="key__shift-caps-lock">|</div>
+          </div>
+        </div>
+      </div>
+      <div class="row">
+        <div class="key key--expand key--word-left" data-key-code="CapsLock">caps lock</div>
+        <div class="key" data-key-code="KeyA">
+          <div class="key__ru">
+            <div class="key__default">ф</div>
+            <div class="key__shift">Ф</div>
+            <div class="key__caps-lock">Ф</div>
+            <div class="key__shift-caps-lock">ф</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">a</div>
+            <div class="key__shift">A</div>
+            <div class="key__caps-lock">A</div>
+            <div class="key__shift-caps-lock">a</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyS">
+          <div class="key__ru">
+            <div class="key__default">ы</div>
+            <div class="key__shift">Ы</div>
+            <div class="key__caps-lock">Ы</div>
+            <div class="key__shift-caps-lock">ы</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">s</div>
+            <div class="key__shift">S</div>
+            <div class="key__caps-lock">S</div>
+            <div class="key__shift-caps-lock">s</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyD">
+          <div class="key__ru">
+            <div class="key__default">в</div>
+            <div class="key__shift">В</div>
+            <div class="key__caps-lock">В</div>
+            <div class="key__shift-caps-lock">в</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">d</div>
+            <div class="key__shift">D</div>
+            <div class="key__caps-lock">D</div>
+            <div class="key__shift-caps-lock">d</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyF">
+          <div class="key__ru">
+            <div class="key__default">а</div>
+            <div class="key__shift">А</div>
+            <div class="key__caps-lock">А</div>
+            <div class="key__shift-caps-lock">а</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">f</div>
+            <div class="key__shift">F</div>
+            <div class="key__caps-lock">F</div>
+            <div class="key__shift-caps-lock">f</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyG">
+          <div class="key__ru">
+            <div class="key__default">п</div>
+            <div class="key__shift">П</div>
+            <div class="key__caps-lock">П</div>
+            <div class="key__shift-caps-lock">п</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">g</div>
+            <div class="key__shift">G</div>
+            <div class="key__caps-lock">G</div>
+            <div class="key__shift-caps-lock">g</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyH">
+          <div class="key__ru">
+            <div class="key__default">р</div>
+            <div class="key__shift">Р</div>
+            <div class="key__caps-lock">Р</div>
+            <div class="key__shift-caps-lock">р</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">h</div>
+            <div class="key__shift">H</div>
+            <div class="key__caps-lock">H</div>
+            <div class="key__shift-caps-lock">h</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyJ">
+          <div class="key__ru">
+            <div class="key__default">о</div>
+            <div class="key__shift">О</div>
+            <div class="key__caps-lock">О</div>
+            <div class="key__shift-caps-lock">о</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">j</div>
+            <div class="key__shift">J</div>
+            <div class="key__caps-lock">J</div>
+            <div class="key__shift-caps-lock">j</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyK">
+          <div class="key__ru">
+            <div class="key__default">л</div>
+            <div class="key__shift">Л</div>
+            <div class="key__caps-lock">Л</div>
+            <div class="key__shift-caps-lock">л</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">k</div>
+            <div class="key__shift">K</div>
+            <div class="key__caps-lock">K</div>
+            <div class="key__shift-caps-lock">k</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyL">
+          <div class="key__ru">
+            <div class="key__default">д</div>
+            <div class="key__shift">Д</div>
+            <div class="key__caps-lock">Д</div>
+            <div class="key__shift-caps-lock">д</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">l</div>
+            <div class="key__shift">L</div>
+            <div class="key__caps-lock">L</div>
+            <div class="key__shift-caps-lock">l</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Semicolon">
+          <div class="key__ru">
+            <div class="key__default">ж</div>
+            <div class="key__shift">Ж</div>
+            <div class="key__caps-lock">Ж</div>
+            <div class="key__shift-caps-lock">ж</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">;</div>
+            <div class="key__shift">:</div>
+            <div class="key__caps-lock">;</div>
+            <div class="key__shift-caps-lock">:</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Quote">
+          <div class="key__ru">
+            <div class="key__default">э</div>
+            <div class="key__shift">Э</div>
+            <div class="key__caps-lock">Э</div>
+            <div class="key__shift-caps-lock">э</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">'</div>
+            <div class="key__shift">"</div>
+            <div class="key__caps-lock">'</div>
+            <div class="key__shift-caps-lock">"</div>
+          </div>
+        </div>
+        <div class="key key--expand key--word-right" data-key-code="Enter">enter</div>
+      </div>
+      <div class="row">
+        <div class="key key--expand key--word-left" data-key-code="ShiftLeft">shift</div>
+        <div class="key" data-key-code="KeyZ">
+          <div class="key__ru">
+            <div class="key__default">я</div>
+            <div class="key__shift">Я</div>
+            <div class="key__caps-lock">Я</div>
+            <div class="key__shift-caps-lock">я</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">z</div>
+            <div class="key__shift">Z</div>
+            <div class="key__caps-lock">Z</div>
+            <div class="key__shift-caps-lock">z</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyX">
+          <div class="key__ru">
+            <div class="key__default">ч</div>
+            <div class="key__shift">Ч</div>
+            <div class="key__caps-lock">Ч</div>
+            <div class="key__shift-caps-lock">ч</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">x</div>
+            <div class="key__shift">X</div>
+            <div class="key__caps-lock">X</div>
+            <div class="key__shift-caps-lock">x</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyC">
+          <div class="key__ru">
+            <div class="key__default">с</div>
+            <div class="key__shift">С</div>
+            <div class="key__caps-lock">С</div>
+            <div class="key__shift-caps-lock">с</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">c</div>
+            <div class="key__shift">C</div>
+            <div class="key__caps-lock">C</div>
+            <div class="key__shift-caps-lock">c</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyV">
+          <div class="key__ru">
+            <div class="key__default">м</div>
+            <div class="key__shift">М</div>
+            <div class="key__caps-lock">М</div>
+            <div class="key__shift-caps-lock">м</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">v</div>
+            <div class="key__shift">V</div>
+            <div class="key__caps-lock">V</div>
+            <div class="key__shift-caps-lock">v</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyB">
+          <div class="key__ru">
+            <div class="key__default">и</div>
+            <div class="key__shift">И</div>
+            <div class="key__caps-lock">И</div>
+            <div class="key__shift-caps-lock">и</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">b</div>
+            <div class="key__shift">B</div>
+            <div class="key__caps-lock">B</div>
+            <div class="key__shift-caps-lock">b</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyN">
+          <div class="key__ru">
+            <div class="key__default">т</div>
+            <div class="key__shift">Т</div>
+            <div class="key__caps-lock">Т</div>
+            <div class="key__shift-caps-lock">т</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">n</div>
+            <div class="key__shift">N</div>
+            <div class="key__caps-lock">N</div>
+            <div class="key__shift-caps-lock">n</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="KeyM">
+          <div class="key__ru">
+            <div class="key__default">ь</div>
+            <div class="key__shift">Ь</div>
+            <div class="key__caps-lock">Ь</div>
+            <div class="key__shift-caps-lock">ь</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">m</div>
+            <div class="key__shift">M</div>
+            <div class="key__caps-lock">M</div>
+            <div class="key__shift-caps-lock">m</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Comma">
+          <div class="key__ru">
+            <div class="key__default">б</div>
+            <div class="key__shift">Б</div>
+            <div class="key__caps-lock">Б</div>
+            <div class="key__shift-caps-lock">б</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">,</div>
+            <div class="key__shift">&lt;</div>
+            <div class="key__caps-lock">,</div>
+            <div class="key__shift-caps-lock">&lt;</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Period">
+          <div class="key__ru">
+            <div class="key__default">ю</div>
+            <div class="key__shift">Ю</div>
+            <div class="key__caps-lock">Ю</div>
+            <div class="key__shift-caps-lock">ю</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">.</div>
+            <div class="key__shift">&gt;</div>
+            <div class="key__caps-lock">.</div>
+            <div class="key__shift-caps-lock">&gt;</div>
+          </div>
+        </div>
+        <div class="key" data-key-code="Slash">
+          <div class="key__ru">
+            <div class="key__default">.</div>
+            <div class="key__shift">,</div>
+            <div class="key__caps-lock">.</div>
+            <div class="key__shift-caps-lock">,</div>
+          </div>
+          <div class="key__en">
+            <div class="key__default">/</div>
+            <div class="key__shift">?</div>
+            <div class="key__caps-lock">/</div>
+            <div class="key__shift-caps-lock">?</div>
+          </div>
+        </div>
+        <div class="key key--expand key--word-right" data-key-code="ShiftRight">shift</div>
+      </div>
+      <div class="row">
+        <div class="key key--ctrl key--word" data-key-code="ControlLeft">ctrl</div>
+        <div class="key key--word" data-key-code="MetaLeft">win</div>
+        <div class="key key--word" data-key-code="AltLeft">alt</div>
+        <div class="key key--expand" data-key-code="Space"></div>
+        <div class="key key--ctrl key--word" data-key-code="ControlRight">ctrl</div>
+        <div class="key key--word" data-key-code="AltRight">alt</div>
+        <div class="key" data-key-code="ArrowLeft">◀</div>
+        <div class="double-key">
+          <div class="double-key__sub-key key" data-key-code="ArrowUp">▲</div>
+          <div class="double-key__sub-key key" data-key-code="ArrowDown">▼</div>
+        </div>
+        <div class="key" data-key-code="ArrowRight">▶</div>
+      </div>
+    </div>
+  `;
+};
+
